Added qemu targets
Debugging is a pain. To try and solve the IRQ bug I'm currently facing
I'm trying to get qemu and gdb to work. They seem to be, the trouble is
qemu's emulation of an arm1176 pagefaults the moment paging is enabled.
It looks like this is going to be tricky.

My real hardware works better than my virtual hardware.